+++ RELEASES.md
@@ -6,12 +6,15 @@
 
 * `structure/dataframe.rs`
 
+### New dependency
+
+* `indexmap`
+
 ### Implement level for DataFrame
 
 - [x] Pretty print
-- [ ] Able to print different length
-- [x] Convert to Row matrix
-- [ ] Convert to Col matrix
+- [x] Able to print different length
+- [x] Convert to Col matrix
 - [x] Call by header (`Index` with header)
 - [ ] Call by row
 - [ ] `IndexMut` with header